Updating experiment scripts

+++ scripts/run_experiment_script.sh
@@ -7,38 +7,55 @@ DEBLOAT_APP="${PWD}/reachability-analysis-1.0-jar-with-dependencies.jar"
 ORIGINAL_SIZE_FILE="${PWD}/original_size.csv"
 DEBLOAT_SIZE_FILE="${PWD}/debloat_size.csv"
 METHOD_DATA_FILE="${PWD}/method_data.csv"
+TEST_DATA_FILE="${PWD}/test_data.csv"
 JAVA="/Library/Java/JavaVirtualMachines/jdk1.8.0_181.jdk/Contents/Home/bin/java"
-
-TEST_SCRIPT="mvn surefire:test"
+POM_EDITOR_APP="${PWD}/pom-editor-1.0-jar-with-dependencies.jar"
+TEST_SCRIPT="${PWD}/test_script.sh"
 
 if [ -f ${ORIGINAL_SIZE_FILE} ]; then
 	>&2 echo "The original size file ("${ORIGINAL_SIZE_DIR}") already exists"
-	echo 1
+	exit 1
 fi
 
 if [ -f ${DEBLOAT_SIZE_FILE} ]; then
 	>&2 echo "The debloat size file ("${DEBLOAT_SIZE_FILE}") already exists"
-	echo 1
+	exit 1
 fi
 
 if [ -f ${METHOD_DATA_FILE} ]; then
 	>&2 echo "The method removal data file ("${METHOD_DATA_FILE}") already exists"
-	echo 1
+	exit 1
 fi
 
-echo "project,is_lib,jar,size_in_kB" >${ORIGINAL_SIZE_FILE}
-echo "project,using_public_entry,using_main_entry,using_test_entry,custom_entry,is_app_prune,is_lib,jar,size_in_kB" >${DEBLOAT_SIZE_FILE}
+if [ -f ${TEST_DATA_FILE} ]; then
+	>&2 echo "The test data file ("${TEST_DATA_FILE}") already exists"
+	exit 1
+fi
+
+echo "project,is_lib,jar,size_in_kB,number_tests_run,number_tests_failed,number_test errors" >${ORIGINAL_SIZE_FILE}
+echo "project,using_public_entry,using_main_entry,using_test_entry,custom_entry,is_app_prune,is_lib,jar,size_in_kB,number_tests_run,number_tests_failed,number_test_errors" >${DEBLOAT_SIZE_FILE}
 echo "project,using_public_entry,using_main_entry,using_test_entry,custom_entry,is_app_prune,lib_methods,app_methods,lib_methods_removed,app_methods_removed" >${METHOD_DATA_FILE}
 
 echo "Setting up 'reachability-analysis' tool"
 if [ ! -f "${DEBLOAT_APP}" ]; then
-	mvn -f ../reachability-analysis/pom.xml clean compile assembly:single >/dev/null
+	mvn -f ../reachability-analysis/pom.xml clean compile assembly:single 2>&1 >/dev/null
 	exit_status=$?
 	if [[ ${exit_status} != 0 ]]; then
-		echo "Cannot build reachability analysis tool"
+		echo "Cannot build 'reachability analysis' tool"
 		exit 1
 	fi
 	cp "../reachability-analysis/target/reachability-analysis-1.0-jar-with-dependencies.jar" .
+fi
+
+echo "Setting up 'pom-editor' tool"
+if [ ! -f "${POM_EDITOR_APP}" ]; then
+	mvn -f ../pom-editor/pom.xml clean compile assembly:single 2>&1 >/dev/null
+	exit_status=$?
+	if [[ ${exit_status} != 0 ]]; then
+		echo "Cannot build 'pom-editor' tool"
+		exit 1
+	fi
+	cp "../pom-editor/target/pom-editor-1.0-jar-with-dependencies.jar" .
 fi
 
 cat ${WORK_LIST} |  while read item; do
@@ -47,10 +64,15 @@ cat ${WORK_LIST} |  while read item; do
 
 	echo "Processing : "${item}
 
-	mvn clean prepare-package dependency:copy-dependencies -DincludeScope=system -DoutputDirectory=target/lib  2>&1 >/dev/null 
+	mvn clean prepare-package dependency:copy-dependencies -DskipTests=true \
+		-DincludeScope=system -DoutputDirectory=target/lib 2>&1 >/dev/null
 	exit_status=$?
 	if [[ ${exit_status} == 0 ]]; then
-		mvn jar:jar dependency:copy-dependencies -DincludeScope=runtime -DoutputDirectory=target/lib 2>&1 >/dev/null
+		
+		#This is necessary to cet the dependencies in the 'runtime' scope 
+		#(we got those in the 'system' scope during the above 'prepare-package')
+		mvn dependency:copy-dependencies -DincludeScope=runtime \
+			-DoutputDirectory=target/lib 2>&1 >/dev/null	
 
 		lib_path=""
 		#Get the size of the library jars
@@ -63,7 +85,7 @@ cat ${WORK_LIST} |  while read item; do
 				unzip "${lib}" -d "target/lib/${lib}" 2>&1 >/dev/null
 				rm "${lib}"		
 
-				echo ${item},1,${lib},$(du -sk "target/lib/${lib}" | cut -f1) >>${ORIGINAL_SIZE_FILE}
+				echo ${item},1,${lib},$(du -sk "target/lib/${lib}" | cut -f1),,, >>${ORIGINAL_SIZE_FILE}
 				if [[ ${lib_path} != "" ]]; then
 					lib_path+=":"
 				fi
@@ -72,7 +94,8 @@ cat ${WORK_LIST} |  while read item; do
 		fi
 
 		#Get the size of the application
-		echo ${item},0,APP,$(du -sk "target/classes" | cut -f1) >>${ORIGINAL_SIZE_FILE}
+		echo ${item},0,APP,$(du -sk "target/classes" | cut -f1),$(${TEST_SCRIPT}\
+		   	"${item_dir}/pom.xml") >>${ORIGINAL_SIZE_FILE}
 
 		temp_file=$(mktemp /tmp/XXXX)
 	
@@ -91,24 +114,35 @@ cat ${WORK_LIST} |  while read item; do
 			lib_classpath="--lib-classpath ${lib_path}"
 		fi
 
-		${JAVA} -jar ${DEBLOAT_APP} ${app_classpath} ${lib_classpath} ${test_classpath} --public-entry --prune-app 2>&1 >$temp_file 
+		${JAVA} -Xms10240m -jar ${DEBLOAT_APP} ${app_classpath} ${lib_classpath} ${test_classpath} \
+			--public-entry --prune-app --remove-methods 2>&1 >$temp_file 
 		exit_status=$?
 		if [[ ${exit_status} == 0 ]];then
 
 			#Get the size of the library sizes
 			ls target/lib | while read lib; do
-				echo ${item},0,1,0,,1,1,${lib},$(du -sk "target/lib/${lib}" | cut -f1) >>${DEBLOAT_SIZE_FILE}
+				echo ${item},0,1,0,,1,1,${lib},$(du -sk "target/lib/${lib}" \
+					| cut -f1),,, >>${DEBLOAT_SIZE_FILE}
+				
+				cd "target/lib/${lib}"
+				temp_jar=$(mktemp /tmp/XXXX)
+				jar cf0 ${temp_jar} ./*
+				cd ../../..
+				rm -r "target/lib/${lib}"
+				mv "${temp_jar}" "target/lib/${lib}"
 			done
 
 			#Get the application size
-			echo ${item},0,1,0,,1,0,APP,$(du -sk "target/classes" | cut -f1) >>${DEBLOAT_SIZE_FILE}
+			${JAVA} -jar ${POM_EDITOR_APP} "${item_dir}/pom.xml" > "${item_dir}/test_pom.xml"
+			echo ${item},0,1,0,,1,0,APP,$(du -sk "target/classes" | cut -f1),$(${TEST_SCRIPT} "${item_dir}/test_pom.xml") >>${DEBLOAT_SIZE_FILE}
 
 			#Get the number of methods/methods wiped for main class as an entry point
 			lib_methods=$(cat ${temp_file} | awk -F, '($1=="number_lib_methods"){print $2}')
 			app_methods=$(cat ${temp_file} | awk -F, '($1=="number_app_methods"){print $2}')
 			lib_methods_removed=$(cat ${temp_file} | awk -F, '($1=="number_lib_methods_removed"){print $2}')
 			app_methods_removed=$(cat ${temp_file} | awk -F, '($1=="number_app_methods_removed"){print $2}')
-			echo ${item},0,1,0,,1,${lib_methods},${app_methods},${lib_methods_removed},${app_methods_removed} >>${METHOD_DATA_FILE}
+			echo ${item},0,1,0,,1,${lib_methods},${app_methods},${lib_methods_removed},\
+				${app_methods_removed},,, >>${METHOD_DATA_FILE}
 		else
 			echo "Could not properly process "${item}
 			echo "Output the following: "
